wip: refactoring continues; - cleaned up api response processing - fixed some initialization runtime errors
