+++ client/cosmos/api/auth.go
@@ -7,6 +7,56 @@ import (
 	"github.com/informalsystems/stakooler/client/cosmos/model"
 )
 
+type AcctResponse struct {
+	Account struct {
+		Type               string `json:"@type"`
+		BaseVestingAccount struct {
+			BaseAccount struct {
+				Address       string `json:"address,omitempty"`
+				PubKey        string `json:"public_key,omitempty"`
+				AccountNumber string `json:"account_number,omitempty"`
+				Sequence      string `json:"sequence,omitempty"`
+			}
+			OriginalVesting []struct {
+				Denom  string `json:"denom"`
+				Amount string `json:"amount"`
+			} `json:"original_vesting"`
+			DelegatedFree []struct {
+				Denom  string `json:"denom"`
+				Amount string `json:"amount"`
+			} `json:"delegated_free"`
+			DelegatedVesting []struct {
+				Denom  string `json:"denom"`
+				Amount string `json:"amount"`
+			} `json:"delegated_vesting"`
+			EndTime string `json:"end_time"`
+		} `json:"base_vesting_account"`
+		StartTime      string `json:"start_time"`
+		VestingPeriods []struct {
+			Length string `json:"length"`
+			Amount []struct {
+				Denom  string `json:"denom"`
+				Amount string `json:"amount"`
+			} `json:"amount"`
+		} `json:"vesting_periods"`
+	} `json:"account"`
+}
+
+func (a AcctResponse) GetBalances() map[int]map[string]string {
+	balances := make(map[int]map[string]string)
+	balances[OriginalVesting] = make(map[string]string)
+	balances[DelegatedVesting] = make(map[string]string)
+
+	for _, balance := range a.Account.BaseVestingAccount.OriginalVesting {
+		balances[OriginalVesting][balance.Denom] = balance.Amount
+	}
+
+	for _, balance := range a.Account.BaseVestingAccount.DelegatedVesting {
+		balances[DelegatedVesting][balance.Denom] = balance.Amount
+	}
+	return balances
+}
+
 func GetPrefix(endpointURL string, client *http.Client) (response model.Bech32PrefixResponse, err error) {
 	var body []byte
 
@@ -23,7 +73,7 @@ func GetPrefix(endpointURL string, client *http.Client) (response model.Bech32Pr
 	return
 }
 
-func GetAccount(address string, endpoint string, client *http.Client) (response *model.AcctResponse, err error) {
+func GetAccount(address string, endpoint string, client *http.Client) (response *AcctResponse, err error) {
 	var body []byte
 
 	url := endpoint + "/cosmos/auth/v1beta1/accounts/" + address
@@ -32,7 +82,7 @@ func GetAccount(address string, endpoint string, client *http.Client) (response 
 		return
 	}
 
-	response = &model.AcctResponse{}
+	response = &AcctResponse{}
 	err = json.Unmarshal(body, response)
 	if err != nil {
 		return
