--- client/cosmos/model/bank.go
@@ -1,25 +1,0 @@
-package model
-
-type BalancesResponse struct {
-	Balances []struct {
-		Denom  string `json:"denom"`
-		Amount string `json:"amount"`
-	} `json:"balances"`
-	Pagination struct {
-		NextKey interface{} `json:"next_key"`
-		Total   string      `json:"total"`
-	} `json:"pagination"`
-}
-
-type DenomMetadataResponse struct {
-	Metadata struct {
-		Description string `json:"description"`
-		DenomUnits  []struct {
-			Denom    string   `json:"denom"`
-			Exponent int      `json:"exponent"`
-			Aliases  []string `json:"aliases"`
-		} `json:"denom_units"`
-		Base    string `json:"base"`
-		Display string `json:"display"`
-	} `json:"metadata"`
-}
